requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
pix2tex>=0.1.1
sympy>=1.12
antlr4-python3-runtime==4.11.1
easyocr>=1.7.1
opencv-python-headless>=4.9.0.80
